Improve set page config test
