+++ e2e_playwright/st_set_page_config_test.py
@@ -11,10 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import re
 
 from playwright.sync_api import Page, expect
 
-from e2e_playwright.conftest import ImageCompareFunction, wait_for_app_run
+from e2e_playwright.shared.app_utils import click_button, expect_exception
 
 
 def test_sets_page_favicon(app: Page):
@@ -39,31 +40,34 @@ def test_sets_page_in_wide_mode(app: Page):
     expect(app_view_container).to_have_attribute("data-layout", "wide")
 
 
-def test_displays_in_wide_mode(app: Page, assert_snapshot: ImageCompareFunction):
-    app_view_container = app.get_by_test_id("stAppViewContainer")
-    assert_snapshot(app_view_container, name="wide-mode")
-
-
 def test_double_setting_set_page_config(app: Page):
-    # Rerun the script to ensure a fresh slate
-    app.keyboard.press("r")
-    wait_for_app_run(app)
-
-    # Test: should not display an error when st.set_page_config is used after an st.* command in a callback
-    app.get_by_test_id("stButton").nth(1).click()
+    # Test: should not display an error when st.set_page_config is used after an st.*
+    # command in a callback
+    click_button(app, "Balloons")
     expect(app.get_by_test_id("stException")).not_to_be_visible()
     expect(app).to_have_title("Heya, world?")
 
-    # Test: should display an error when st.set_page_config is called multiple times in a callback
-    app.get_by_test_id("stButton").nth(2).click()
-    expect(app.get_by_test_id("stException")).to_contain_text(
-        "set_page_config() can only be called once per app page"
-    )
+    # Test: should display an error when st.set_page_config is called
+    # multiple times in a callback
+    click_button(app, "Double Set Page Config")
+    expect_exception(app, "set_page_config() can only be called once per app page")
     expect(app).to_have_title("Change 1")
 
-    # Test: should display an error when st.set_page_config is called after being called in a callback
-    app.get_by_test_id("stButton").nth(3).click()
-    expect(app.get_by_test_id("stException")).to_contain_text(
-        "set_page_config() can only be called once per app page"
-    )
+    # Test: should display an error when st.set_page_config is called after
+    # being called in a callback
+    click_button(app, "Single Set Page Config")
+    expect_exception(app, "set_page_config() can only be called once per app page")
+
     expect(app).to_have_title("Change 3")
+
+
+def test_st_set_page_config_sets_page_icon(app: Page):
+    click_button(app, "Page Config With Local Icon")
+
+    expect(app).to_have_title("With Local Icon")
+    favicon_element = app.locator("link[rel='shortcut icon']")
+    expect(favicon_element).to_have_count(1)
+    expect(favicon_element).to_have_attribute(
+        "href",
+        re.compile(r"d1e92a291d26c1e0cb9b316a93c929b3be15899677ef3bc6e3bf3573\.png"),
+    )
